moved private data into protected folder
