+++ routes/api/resume.js
@@ -1,7 +1,7 @@
-import courses from '../../app/data/courses';
-import degrees from '../../app/data/degrees';
-import positions from '../../app/data/positions';
-import { skills, categories } from '../../app/data/skills';
+import courses from '../../app/data/protected/courses';
+import degrees from '../../app/data/protected/degrees';
+import positions from '../../app/data/protected/positions';
+import { skills, categories } from '../../app/data/protected/skills';
 
 export default (req, res) => {
   res.json({
